Move `electCommitte` and `electBasicCommittee` to conway `ImpTest`
for easier reuse across tests
