+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EpochSpec.hs
@@ -9,7 +9,7 @@
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module Test.Cardano.Ledger.Conway.Imp.EpochSpec (spec, electBasicCommittee) where
+module Test.Cardano.Ledger.Conway.Imp.EpochSpec (spec) where
 
 import Cardano.Ledger.BaseTypes (textToUrl)
 import Cardano.Ledger.Coin
@@ -20,33 +20,22 @@ import Cardano.Ledger.Conway.Governance (
   GovAction (..),
   ProposalProcedure (..),
   Voter (..),
-  cgEnactStateL,
-  ensCommitteeL,
  )
 import Cardano.Ledger.Conway.PParams (
-  ppCommitteeMaxTermLengthL,
-  ppDRepVotingThresholdsL,
   ppGovActionDepositL,
   ppGovActionLifetimeL,
  )
-import Cardano.Ledger.Credential (Credential (..))
-import Cardano.Ledger.Keys
 import Cardano.Ledger.Shelley.LedgerState (
   asTreasuryL,
   esAccountStateL,
-  esLStateL,
-  lsUTxOStateL,
   nesEsL,
-  utxosGovStateL,
  )
 import Cardano.Ledger.Val
 import Data.Default.Class (Default (..))
-import qualified Data.Map.Strict as Map
 import Data.Maybe (fromJust)
-import Data.Maybe.Strict (StrictMaybe (..), isSJust)
+import Data.Maybe.Strict (StrictMaybe (..))
 import Lens.Micro ((&), (.~))
 import Test.Cardano.Ledger.Conway.ImpTest
-import Test.Cardano.Ledger.Core.Rational (IsRatio (..))
 import Test.Cardano.Ledger.Imp.Common
 
 spec ::
@@ -152,60 +141,3 @@ spec =
       expectMissingGovActionId govActionId
 
       getRewardAccountAmount rewardAcount `shouldReturn` deposit
-
-electBasicCommittee ::
-  forall era.
-  ( HasCallStack
-  , ConwayEraImp era
-  , GovState era ~ ConwayGovState era
-  ) =>
-  ImpTestM era (Credential 'DRepRole (EraCrypto era), Credential 'HotCommitteeRole (EraCrypto era))
-electBasicCommittee = do
-  logEntry "Setting up PParams and DRep"
-  modifyPParams $ \pp ->
-    pp
-      & ppDRepVotingThresholdsL
-        .~ def
-          { dvtCommitteeNormal = 1 %! 1
-          , dvtCommitteeNoConfidence = 1 %! 2
-          , dvtUpdateToConstitution = 1 %! 2
-          }
-      & ppCommitteeMaxTermLengthL .~ 10
-      & ppGovActionLifetimeL .~ 2
-      & ppGovActionDepositL .~ Coin 123
-  khDRep <- setupSingleDRep
-
-  logEntry "Registering committee member"
-  khCommitteeMember <- freshKeyHash
-  let
-    committeeAction =
-      UpdateCommittee
-        SNothing
-        mempty
-        (Map.singleton (KeyHashObj khCommitteeMember) 10)
-        (1 %! 2)
-  gaidCommitteeProp <- submitGovAction committeeAction
-
-  submitYesVote_ (DRepVoter $ KeyHashObj khDRep) gaidCommitteeProp
-
-  let
-    assertNoCommittee = do
-      committee <-
-        getsNES $
-          nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . cgEnactStateL . ensCommitteeL
-      impAnn "There should not be a committee" $ committee `shouldBe` SNothing
-  logRatificationChecks gaidCommitteeProp
-  assertNoCommittee
-
-  passEpoch
-  logRatificationChecks gaidCommitteeProp
-  assertNoCommittee
-  passEpoch
-  do
-    committee <-
-      getsNES $
-        nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . cgEnactStateL . ensCommitteeL
-    impAnn "There should be a committee" $ committee `shouldSatisfy` isSJust
-
-  khCommitteeMemberHot <- registerCommitteeHotKey khCommitteeMember
-  pure (KeyHashObj khDRep, KeyHashObj khCommitteeMemberHot)
