+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/State/Imp/QuerySpec.hs
@@ -20,10 +20,6 @@ import Cardano.Ledger.Conway.Governance (
   Committee (..),
   ConwayGovState,
   EraGov (..),
-  GovAction (..),
-  GovActionPurpose (..),
-  PrevGovActionId (..),
-  Voter (DRepVoter),
   cgEnactStateL,
   ensCommitteeL,
  )
@@ -277,28 +273,6 @@ spec =
       ImpTestM era ()
     expectNoFilterQueryResult =
       expectQueryResult mempty mempty mempty
-
-electCommittee ::
-  forall era.
-  ( HasCallStack
-  , ConwayEraImp era
-  ) =>
-  StrictMaybe (PrevGovActionId 'CommitteePurpose (EraCrypto era)) ->
-  KeyHash 'DRepRole (EraCrypto era) ->
-  Set.Set (KeyHash 'ColdCommitteeRole (EraCrypto era)) ->
-  Map.Map (KeyHash 'ColdCommitteeRole (EraCrypto era)) EpochNo ->
-  ImpTestM era (PrevGovActionId 'CommitteePurpose (EraCrypto era))
-electCommittee prevGovId drep toRemove toAdd = do
-  let
-    committeeAction =
-      UpdateCommittee
-        prevGovId
-        (Set.map KeyHashObj toRemove)
-        (Map.mapKeys KeyHashObj toAdd)
-        (1 %! 2)
-  gaidCommitteeProp <- submitGovAction committeeAction
-  submitYesVote_ (DRepVoter $ KeyHashObj drep) gaidCommitteeProp
-  pure (PrevGovActionId gaidCommitteeProp)
 
 setPParams ::
   forall era.
